create narrative module and separate out seams; fix imports so local tests pass
