+++ refiner/app/services/ecr/section/generic_matching.py
@@ -14,7 +14,7 @@ from ..model import (
     SectionRunResult,
     SectionSpecification,
 )
-from .narrative import (
+from ..narrative import (
     create_minimal_section,
     remove_all_comments,
     replace_narrative_with_removal_notice,
